Refactor, use pg with docker container, and add crud db operations for each db.

+++ core/src/timing.rs
@@ -5,9 +5,9 @@ pub const DAY: u64 = HOUR * 24;
 pub const HOUR: u64 = MINUTE * 60;
 pub const MINUTE: u64 = 60;
 
-pub fn get_timestamp() -> u64 {
+pub fn get_timestamp() -> i64 {
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap()
-        .as_secs()
+        .as_secs() as i64
 }
